celeryconfig/requirements.txt
slackclient
python-dotenv
celery
redis
